Home
Components
Storybook
Documentation
Haddock

Hypered.Html

Description

Define HTML code for the Hypered design system.

Documentation

generate :: FilePath -> Text -> (FilePath -> Html) -> IO ()

Generate both the normal and pretty-printed HTML versions.

generateHtml :: Config -> FilePath -> FilePath -> Text -> Html -> IO ()

prettyHtml :: Config -> FilePath -> FilePath -> Text -> Html -> IO ()

partialHtml :: Config -> FilePath -> FilePath -> Text -> Html -> IO ()

Same as prettyHtml but doesn't wrap the content to create a full standalone HTML document.

data Font

Constructors

IbmPlex

Inter

Font String

fontClass :: Font -> String

fontCss :: Font -> String

data Config

Constructors

Config

Fields

cStaticPath :: FilePath

cFont :: Font

defaultConfig :: Config

mkRelativize :: FilePath -> FilePath -> FilePath

document :: Config -> FilePath -> Text -> Html -> Html

nav :: Html -> Html

The main horizontal navigation component. It uses "flex justify-between" so it is better to group its contained links with divs. With a single div, they are all grouped on the left. With two, the first is grouped to the left and the second to the right. See https://hypered.github.io/design-system/storybook/?path=/story/navigation--navigation for examples.

navigation :: FilePath -> Html

Horizontal navigation at the top of a page.

navigationNoteed :: Html

Horizontal navigation at the top of a page, at the same level as main wrapper and footer.

navigationNoteed' :: Html

Same as navigationNoteed but with links on the right, except the first one.

navigationTemplate :: Html

wrapPost :: Html -> MarkupM () -> Html

Content wrapper, for a blog post, at the same level as navigation and footer.

wrap :: Html -> Html

The main content wrapper, at the same level as navigation.

footer :: Html -> Html

The footer, at the same level as both navigation and wrap.

section :: Html -> Html

The main content, as a left column.

aside :: Html

A right column, with a title and a list of links.

bannerGreen :: Html -> Html

Green variant of a banner.

bannerYellow :: Html -> Html

Yellow variant of a banner.

bannerRed :: Html -> Html

Red variant of a banner.

banner :: [Char] -> Html -> Html

buttonPrimary :: Html -> Html

buttonPrimaryDisabled :: Html -> Html

buttonSecondary :: Html -> Html

buttonSecondaryDisabled :: Html -> Html

buttonFullWidth :: Html -> Html

codeBlock :: Html

title :: Html

sidebarTitle :: Html -> Html

sidebarUL :: Html -> Html

sidebarLI :: Html -> Html

sidebarLink :: Html -> AttributeValue -> Html

sidebar :: (Foldable t, Foldable t1) => t1 (Html, t (Html, AttributeValue)) -> Html

exampleSidebar :: MarkupM ()

exampleSidePanel :: MarkupM ()